cleanup(core): move mapProjectGraphFiles to target project locator
