+++ packages/eslint-plugin-nx/src/rules/enforce-module-boundaries.spec.ts
@@ -1,9 +1,12 @@
 import type { FileData, ProjectGraph } from '@nrwl/devkit';
-import { DependencyType, mapProjectGraphFiles } from '@nrwl/devkit';
+import { DependencyType } from '@nrwl/devkit';
 import * as parser from '@typescript-eslint/parser';
 import { TSESLint } from '@typescript-eslint/utils';
 import { vol } from 'memfs';
-import { TargetProjectLocator } from 'nx/src/utils/target-project-locator';
+import {
+  TargetProjectLocator,
+  mapProjectGraphFiles,
+} from 'nx/src/utils/target-project-locator';
 import enforceModuleBoundaries, {
   RULE_NAME as enforceModuleBoundariesRuleName,
 } from './enforce-module-boundaries';
